Modular handle
