+++ lib/express/index.js
@@ -1,7 +1,6 @@
-var http = require('http');
-var url = require('url');
+const http = require('http');
 const parsePaths = require('./parsePaths');
-
+const handler = require('./handler');
 const verbs = ['get', 'post'];
 
 Router = {};
@@ -14,6 +13,9 @@ Router.listen = function(port, host, callback) {
   server.listen(port, host, callback);
 };
 
+// Route requests
+Router.handle = handler;
+
 // Build HTTP verb methods
 verbs.forEach(verb => {
   Router.routes[verb] = Router.routes[verb] || {};
@@ -24,41 +26,5 @@ verbs.forEach(verb => {
     Router.routes[verb][parsedPath.pathRegex.source].params = parsedPath.params;
   };
 });
-
-// Route requests
-Router.handle = function(req, res) {
-  console.log('Routing... ... ...');
-  const verb = req.method.toLowerCase();
-  let path = url.parse(req.url).pathname;
-
-  // Remove trailing slashes
-  if (path.length > 1 && path[path.length - 1] === '/') {
-    path = path.slice(0, path.length - 1);
-  }
-
-  // Iterate over the path regexes for our given verb
-  // testing them against the current path
-  var found = false;
-  for (let routerPath in Router.routes[verb]) {
-    let pathRegex = new RegExp(routerPath);
-    if (pathRegex.test(path)) {
-      var params = {};
-      var registeredParams = Router.routes[verb][routerPath].params;
-      pathSections = path.split('/').slice(1);
-      for (let index in registeredParams) {
-        params[registeredParams[index]] = pathSections[index];
-      }
-      req.params = params;
-      // Path matches! Route away...
-      Router.routes[verb][routerPath](req, res);
-      found = true;
-      break;
-    }
-  }
-  if (!found) {
-    res.statusCode = 404;
-    res.end('404 NOT FOUND, LOSER');
-  }
-};
 
 module.exports = Router;
